Adding HAL semaphore support for statuses-as-failure-payloads. (IREE-18912)
This allows an implementation to have a single atomic value for a
semaphore that encodes the user payload or an error payload that
optionally references an iree_status_t object. Implementations not using
the status feature can ignore it but must perform a
greater-than-or-equal check on `IREE_HAL_SEMAPHORE_FAILURE_VALUE`
instead of equality.