Adding iree_hal_buffer_placement_t info to allocated HAL buffers. (IREE-19160)
This allows users to query for the device, queue affinity, and origin
flags of an allocated buffer so that they can decide whether to copy
buffers when moving across devices/queues, perform synchronous or
asynchronous deallocations, and select queues to perform such
operations. The downside is that prior to this buffers could be defined
as shared across multiple devices that share a single allocator -
because it's still the case that they can be shared "placement" was used
as it's not "availability" or "access permissions" but only indicating
the origin of the buffer and where it's preferred location is. Buffer
compatibility queries are still the way to determine access visibility.

The base `iree_hal_buffer_t` structure was cleaned up a bit and though
still ugly is better prepared for removal of the device allocator back
reference. IREE-19159 tracks removing the allocator reference that is
currently only used by the caching allocator due to our lack of dynamic
casts.

Breaking API changes:
* `iree_hal_buffer_subspan` now requires an `iree_allocator_t
host_allocator` that was previously implicit.
* `iree_hal_subspan_buffer_initialize` removed as it was not safe and
shouldn't have been used.
* `iree_hal_deferred_buffer_t` removed as it is not possible to do the
placement checks without an allocated buffer reference. This is used in
a branch for async allocations on CPU but either a different approach is
required there or it can be moved into `iree/hal/local/` where we can
make assertions about universal accessibility.
* `iree_hal_heap_buffer_wrap` and all other buffer creation now requires
a placement.
* `iree_hal_buffer_initialize` used by HAL implementations now requires
a placement.