Fix floating point type conversions when the result is in the denormal range (IREE-20242)
The condition for generating a denormal (which we just flush to zero)
was incorrect. The new condition is very simple, checks that the biased
converted exponent would be <= 0 which is the definition of a denormal.
The added testcases show what is being fixed here. Prior to the fix,
these cases converted to some large values as the exponent wrapped
around!

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>