Adding mimalloc v3 as an optional system allocator. (IREE-20730)
Setting `-DIREE_ALLOCATOR_SYSTEM=mimalloc` will fetch v3 of mimalloc and
switch the default system allocator to using it by statically linking it
into iree::base. v3 was chosen because people seem to say it's
significantly better in serving environments (or anywhere with a lot of
threads) and that's what IREE is designed for.

If users bring their own and set `IREE_USE_SYSTEM_DEPS` they'll get what
they have installed. Alternatively, a user can decide how they want to
integrate mimalloc themselves, have it replace malloc/free globally, and
not bother with any IREE-specific settings. That's the only way to make
mimalloc work as an allocator in the compiler (outside of consteval) as
well, if we wanted to do that.

This is not enabled by default as mimalloc is not supported in all
configurations. Release builds could set it as default for the
configurations where it is supported if they want.

Fixes IREE-16812.