Adding support for `#hal.device.optimal<...>` through to runtime. (IREE-20879)
This is an experimental approach for representing a set of device
affinities that are able to be resolved at runtime. Currently this is
limited to allocation-related ops but may be extended in the future to
allow for runtime-provided device rankings based on benchmarked
performance.

The new `#hal.device.optimal<...>` affinity attribute represents a set
of affinities that an operation is able to execute with. The expectation
is that the exact affinity can be resolved during compilation if
sufficient information is available (TBD) or at runtime based on a
user-provided policy (`iree_hal_module_device_policy_t`). As with other
treatments of device topology in the runtime this policy is left to the
hosting framework or application to decide how they want to configure
their overall topology. The common IREE command line tooling adds a
`--device_lead_allocator=N` flag to let the user denote which device
ordinal (of `--device=` flags) is to be chosen when a
`#hal.device.optimal` request makes it all the way to runtime. For
simple cases of CPU+accelerator this allows for hack-free ways to
indicate the lead allocator that is responsible for any buffer
allocation that is used on multiple devices.

The `#hal.device.optimal<...>` attr lowers to a `hal.allocator.select`
op/runtime call that takes a list of devices/queue masks along with the
memory types and buffer usage of the request and returns the selected
device/queue mask. Size is not included such that the selection can be
memoized across an entire program at initialization time. A pass is
added that runs at the same time as device query memoization to hoist
the selection into globals. Future changes can enhance the pass to fold
the selection when sufficient information has been provided.

Future work on IREE-20855 will result in affinity assignment on resource
allocations lowering into this new `#hal.device.optimal<...>` attr and
deallocations switching to either use the recently added `origin` attr
or the same optimal attr (given that it should be consistent and
provides more information). Future work on IREE-20856 will ensure that
allocations that may be used on multiple devices (regardless of whether
runtime-selected) get the appropriate bits set but that may require
IREE-20854.

Progress on IREE-20851.
Progress on IREE-20855 (analysis will lower into this attr).
Progress on IREE-20856 (will be used to set bits).
Fixes IREE-20857.