Allow importing an external stream into HIP. (IREE-20972)
This is for cases where we want a very tight integration with an
external application using HIP. This allows us to share a stream withan
external application. This forces dispatches to be serialized with the
main thread so we aren't submitting to the stream while someone else is
as well.

---------

Signed-off-by: Andrew Woloszyn <andrew.woloszyn@gmail.com>