Adding support for group_any in iree_thread_affinity_t. (IREE-21089)
This allows code launching threads to specify that a thread should be
assigned to a processor associated with a specific group (Windows
GROUP_AFFINITY, Linux NUMA node ID, etc) instead of a specific processor
within the group. This is useful for threads created in various parts of
the codebase that don't globally coordinate (service workers/etc), while
things like the task system that explicitly layout large collections of
threads will continue to explicitly assign them.