Adding device-side tracing macros and a device-side trace buffer. (IREE-21046)
Host code is required to allocate and parse the trace buffer but that
needs a reworking in future PRs and is excluded here. The
`iree_hal_amdgpu_device_trace_buffer_t` struct and how the ringbuffer
works are going to change a lot, while the trace events and how the
device populates them should not. For now the macros (which are today
always no-op'ed) are useful for future PRs that reference them.

Tracing in on-device C works nearly identically to normal tracing on the
host. The code location adjustment in the HAL executable loader allows
us to translate static source locations embedded in the device binary
(HSACO/ELF) to locations on the host we can pass on to tooling like
Tracy. The `iree_hal_amdgpu_device_trace_buffer_t` is a per-queue
ringbuffer of trace events that is to be read by a translation thread on
the host.

The primary use of this is diagnosing program behavior in our
device-side scheduler code. We don't care about this adapting to
instruction-level tracing as we won't be inspecting our device-side
scheduling code as part of normal execution.

The way the events are defined allows us to route these to Tracy as if
the queue was a host thread (vs a GPU zone with limited features).