Adding device->host service worker. (IREE-21094)
The worker is assigned to nodes nearby the device they are serving and
mimics an HSA/AQL queue. The intent is that HSA-specific tooling can
intercept our packets but I'm not sure there's any _good_ tooling that
does that we care about. Since the difference between an AQL queue and
what we'd do is small I decided to stick to the standard here. In the
future we can evaluate whether it makes more sense to have our own
packet format (instead of agent dispatches) or our own ringbuffers for
common operations like releasing resources or signaling semaphores from
the host.