Adding IREE_ENABLE_RUNTIME_COVERAGE cmake mode. (IREE-21191)
This enables LLVM coverage information for all runtime libraries, test
binaries, and some specific runtime tools. We may want to extend this to
support the compiler in the future so everything has been left
runtime-prefixed. Additionally, a script and some cmake targets are
provided to convert captured coverage data into LCOV info files for
tooling. The implementation focuses on automating coverage on CI (in the
future), producing LCOV reports, and integrating with IDE tooling.

As noted in the documentation this is roughly shaped to fit in with
VSCode's cmake-tools coverage support. Prior to a coverage run triggered
from the UI all profiling information is cleared via the
`iree-runtime-coverage-clear-all` target, cmake-tools runs the chosen
tests and collects new information, and then the LCOV file is exported
via the `iree-runtime-coverage-export` target. Users can build these
targets themselves to handle one-shot coverage or accumulating many
coverage runs from different ctest or manual invocations.

The major point of complexity here is around LLVM not being able to
resolve a .profraw output from a tool to the source objects it was
compiled with. If doing a full build, full test, and then single export
that works ok but if trying to subset it falls over fast due to the
performance of llvm-cov with many inputs: it can take a minute or more
to handle all of the runtime object files if passed together. To make
the common interactive coverage case faster we export a map of CMake
targets to their corresponding output binary and use a naming convention
on the .profraw files to help the export script subset the target list.
Users can also specify their own objects to inspect in addition to the
automatic ones with the `IREE_RUNTIME_COVERAGE_OBJECTS` variable pointed
at one or more .o, .a, or executable binaries.

Example running two tests with coverage and the resulting coverage
tree/inline coverage overlay:

![image](https://github.com/user-attachments/assets/cede4b4e-7dd3-4f0e-87d1-55f9aebb75d2)

![image](https://github.com/user-attachments/assets/ee96e1fb-8604-4769-8e94-d4ff760bdd74)

VSCode workspace configuration:
```json
    "cmake.configureArgs": ["-DIREE_ENABLE_RUNTIME_COVERAGE=ON"],
    "cmake.preRunCoverageTarget": "iree-runtime-coverage-clear-all",
    "cmake.postRunCoverageTarget": "iree-runtime-coverage-export",
    "cmake.coverageInfoFiles": [
      "${command:cmake.buildDirectory}/coverage/runtime.lcov.info"
    ],
```

Beyond basic manual/CI/IDE support everything else is left to future
work (aforementioned compiler coverage, PGO, other export formats like
gcov, other compilers like gcc, etc).