Add e2e test for split reduction using tiling. (IREE-21374)
This adds end to end tests for split-reduction tiling. To facilitate
testing add a new attribute that can be used to specify the reduction
tile sizes for the operations on the input MLIR. This is then picked up
to drive splitting reductions into two dispatches, one for the partial
reduce and one for the final combination. A pipeline test to verify that
this attribute works as intended is added as well.

Delete the command line option that was used to turn on split-reduction
using tiling.

---------

Signed-off-by: MaheshRavishankar <mravisha@amd.com>