Fix string parsing of i8 and i16 cl values (IREE-21409)
Signless integers can be provided as either signed or unsigned, and when
parsing them we first parse into i32. This means that a valid signless
integer can either be within the valid signed or unsigned range for <32
bit types.

Hex values should only verify bitwidth, but this check is only
marginally less restrictive than a bitwidth check. For example, passing
`0xFFFFFFFE` would parse as `-2` and be considered valid when in reality
we would normally want to consider that out of range, but this is likely
not a common case and doesn't throw an error. Not specializing for hex
inputs keeps the implementation simpler.