Adding iree_hal_device_queue_dispatch. (IREE-21630)
This is emulated today by creating a command buffer with a single
dispatch and executing it. HAL implementations that can more efficiently
perform a single dispatch in queue order should do so. We don't want
people doing queue-level fill/copy/dispatch, but they are useful for
glue code and allowing users to avoid the command buffer cost for O(1)
operations is worth it.

Fixes IREE-21629.