Adding semaphore creation and wait flags for controlling behavior. (IREE-21619)
This set of changes adds arguments to semaphore APIs that allow
applications to signal their intended usage of semaphores they create to
implementations (which queues they may be used on, whether they need
interrupt services, and whether they are exportable) and when they wait
what kind of latency they are expecting as a hint for the wait behavior.

No implementations are updated here, all of these flags are for future
work. Behavior is defaulted to what it is today before these additions
(interrupts/exports requested, no special handling of queue affinity,
and blocking waits). This is the most general behavior that should
always be correct with only usage of the new flags potentially creating
problems.

Fixes IREE-21615.