Adding iree_hal_device_queue_host_call and emulation.  (IREE-21653)
This allows for both blocking and non-blocking device->host calls.

Emulation is provided targets that aren't yet using their native
features (CUDA/HIP/Metal) or don't have them (Vulkan), but it should
never be used once we start relying on this for programs as the
performance is terrible. The CPU sync and task implementations are done
here as the emulation is incompatible with sync semantics and it's
possible to implement it on the task system fairly easily.

I split out the existing queue emulation utilities out of the device.c
so high-fidelity backends can eventually not even link that code in. I
added the host call emulation in its own target so we can avoid
introducing threading dependencies into `iree::hal` for the emulation
(and makes it clearer what's part of the API vs what's an implementation
detail).

I suspect there may be some HIP flakes and we can disable the CTS
there/file issues if they pop up. I think there's a few cases in the HIP
semaphore that don't quite work but it'd be better to improve the
semaphore tests first. HIP is using an emulated host call here and
that's pretty much just a thread and some semaphores and should be
possible to test independent of the host call logic.

Fixes IREE-21631.