Adding iree_hal_executable_export_info_t and queries. (IREE-21754)
This adds per entry-point flags and workgroup size to enable upcoming
HAL APIs for querying the same. Exports also now have parameter
reflection information to allow for name -> parameter mapping and
parameter packing from raw parameter lists. A placeholder for occupancy
information is added but not yet specified so it won't require a
breaking change in the future.

This is a breaking local executable library format change and bumps it
to version 0.6.
This could have been done without breaking but in the last revision the
dispatch attrs table was incorrectly grown to 72 bytes/entry and that is
bad (2 cache lines for single dispatch info) and needed to be corrected.
Future runtime changes will require the workgroup size as well and now
that we're emitting it from the compiler the runtime can assume it is
always present.

While changing these APIs I finally fixed the ancient `int32_t
entry_point;`.

A basic CTS test was added for the export info but no targets currently
report full parameter information when produced by the compiler so tests
of that are left for when that lands.

Fixes IREE-21750. The local executable loaders support this API but all
other targets are left to future work.