Adding sysfs topology detection logic and switching to it by default. (IREE-22455)
This avoids using cpuinfo just to get this standardized information and
will let us drop that as a dependency. We'll let this soak for a bit to
see if anyone notices before finally removing cpuinfo entirely. For now
`-DIREE_ENABLE_CPUINFO=ON` can be used to override and cpuinfo will be
fetched (or the system dep will be used).

This is not meant to be the end-all be-all of topology information but
is more featureful than cpuinfo for what we needed and can be easily
updated. Testing is going to be its own thing so I have a way to
manually do it for now. Real testing will require a corpus of sysfs info
in json/yaml that we plug in to our file I/O as a virtual filesystem -
otherwise these sysfs trees grow to 10-20MB. Since we're using standard
sysfs information (no tables/etc builtin) we should be fairly robust
outside of kernel versioning (which are hopefully handled well enough 🤞)
and our own logic, and if we go heavy on CPU workloads again we can
invest in tuning that (everything today is best-guess only).

Progress on IREE-8469.