[runtime] Require aligned memory accesses by default (IREE-22557)
The C/C++ standards require all memory accesses to be aligned, even if
the target hardware allows for unaligned accesses. Change the default to
require alignment, so that we get closer to being UB-warning free when
running with the UBSan enabled.

Advanced users can choose to disable this via the build system.

Also simplify the implementation of unaligned loads/stores and rely on
`memcpy` being a compiler intrinsic.

Issue: IREE-22555

ci-extra: test_torch, windows_x64_msvc